// packages/sdk/src/sdk/api/generated/default/models/CreateCommentRequest.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * API
 * Audius V1 API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface CreateCommentRequest
 */
export interface CreateCommentRequest {
    /**
     * Type of entity being commented on
     * @type {string}
     * @memberof CreateCommentRequest
     */
    entityType: string;
    /**
     * ID of the entity being commented on
     * @type {string}
     * @memberof CreateCommentRequest
     */
    entityId: string;
    /**
     * Comment text
     * @type {string}
     * @memberof CreateCommentRequest
     */
    body: string;
    /**
     * Optional comment ID
     * @type {number}
     * @memberof CreateCommentRequest
     */
    commentId?: number;
    /**
     * Parent comment ID if this is a reply
     * @type {number}
     * @memberof CreateCommentRequest
     */
    parentId?: number;
}

/**
 * Check if a given object implements the CreateCommentRequest interface.
 */
export function instanceOfCreateCommentRequest(value: object): value is CreateCommentRequest {
    let isInstance = true;
    isInstance = isInstance && "entityType" in value && value["entityType"] !== undefined;
    isInstance = isInstance && "entityId" in value && value["entityId"] !== undefined;
    isInstance = isInstance && "body" in value && value["body"] !== undefined;

    return isInstance;
}

export function CreateCommentRequestFromJSON(json: any): CreateCommentRequest {
    return CreateCommentRequestFromJSONTyped(json, false);
}

export function CreateCommentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateCommentRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'entityType': json['entityType'],
        'entityId': json['entityId'],
        'body': json['body'],
        'commentId': !exists(json, 'commentId') ? undefined : json['commentId'],
        'parentId': !exists(json, 'parentId') ? undefined : json['parentId'],
    };
}

export function CreateCommentRequestToJSON(value?: CreateCommentRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'entityType': value.entityType,
        'entityId': value.entityId,
        'body': value.body,
        'commentId': value.commentId,
        'parentId': value.parentId,
    };
}