// packages/sdk/src/sdk/api/generated/default/models/UpdateCommentRequest.ts
/* tslint:disable */
/* eslint-disable */
// @ts-nocheck
/**
 * API
 * Audius V1 API
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UpdateCommentRequest
 */
export interface UpdateCommentRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateCommentRequest
     */
    body: string;
}

/**
 * Check if a given object implements the UpdateCommentRequest interface.
 */
export function instanceOfUpdateCommentRequest(value: object): value is UpdateCommentRequest {
    let isInstance = true;
    isInstance = isInstance && "body" in value && value["body"] !== undefined;

    return isInstance;
}

export function UpdateCommentRequestFromJSON(json: any): UpdateCommentRequest {
    return UpdateCommentRequestFromJSONTyped(json, false);
}

export function UpdateCommentRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): UpdateCommentRequest {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'body': json['body'],
    };
}

export function UpdateCommentRequestToJSON(value?: UpdateCommentRequest | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'body': value.body,
    };
}